+++ src/service/stringConverter/json.ts
@@ -1,10 +1,10 @@
 import { TokenInfo } from "../codeParser";
 import { isStringToken, isUnknownToken } from "../literalParser/interface";
-import { StringConverter, StringConverterConvertResult, StringConverterMatchResult, StringConverterMeta, StringConverterOptions } from "./interface";
+import { StringConverter, StringConverterConvertResult, StringConverterMatchResult, StringConverterMeta, StringConverterMatchOptions } from "./interface";
 
 export class JsonParser implements StringConverter<any> {
     meta: StringConverterMeta = {
-        id: "json-parser",
+        id: "json-formater",
         name: "Format JSON String",
         resultLanguageId: "json",
         specInfo: {
@@ -20,7 +20,7 @@ export class JsonParser implements StringConverter<any> {
         }
     };
 
-    match(tokenInfo: TokenInfo, options?: StringConverterOptions): StringConverterMatchResult<any> {
+    match(tokenInfo: TokenInfo, options?: StringConverterMatchOptions): StringConverterMatchResult<any> {
         if (!isStringToken(tokenInfo.type) && !isUnknownToken(tokenInfo.type)) {
             return { matched: false };
         }
@@ -39,7 +39,7 @@ export class JsonParser implements StringConverter<any> {
         }
     }
 
-    convert(tokenInfo: TokenInfo, byproductOfMatch?: any, options?: StringConverterOptions): StringConverterConvertResult {
+    convert(tokenInfo: TokenInfo, byproductOfMatch?: any, options?: StringConverterMatchOptions): StringConverterConvertResult {
         const jsonObj = byproductOfMatch || JSON.parse(tokenInfo.text);
         return {
             result: JSON.stringify(jsonObj, null, 2),